@@ -52,9 +52,9 @@
     "po2json": "0.4.1",
     "punycode": "2.1.1",
     "querystring": "0.2.0",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-table": "7.3.2",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-table": "7.8.0",
     "redux": "3.6.0",
     "shell-quote": "1.7.3",
     "shelljs": "0.8.5",
@@ -79,7 +79,7 @@
     "eslint-plugin-react": "7.20.6",
     "eslint-plugin-react-hooks": "4.1.2",
     "file-url": "2.0.2",
-    "flow-bin": "0.175.1",
+    "flow-bin": "0.179.0",
     "gettext-parser": "3.1.0",
     "http-proxy": "1.18.1",
     "json5": "2.1.3",
